fastapi-server/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
aspose-words==24.2.0
bidict==0.23.1
brute==0.0.3
click==8.1.7
cobble==0.1.3
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
lolcat==1.4
mammoth==1.7.0
pydantic==2.6.3
pydantic_core==2.16.3
python-engineio==4.9.0
python-multipart==0.0.9
python-socketio==5.11.1
rarfile==4.0
simple-websocket==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.10.0
unrar==0.4
uvicorn==0.27.1
wsproto==1.2.0
